feat: enable latch debugging for tests
Adds a new ENABLE_LATCH_DEBUGGING compile-time flag when enabled will
keep track of the lock and unlock calls for both shared and exclusive
latches. This is done to catch mistakes in code when a lock call does
not have a matching unlock call, or the other way around. When the
tests are run and the SharedLatch destructor is called, if the counts
are not zero, then the tests will fail.